fix: update rate limit configuration to use float for window and global_window

+++ src/config/rate_limit.py
@@ -4,11 +4,11 @@ from pydantic_settings import BaseSettings
 
 class RateLimitEndpoint(BaseModel):
     limit: int = 100
-    window: int = 60
+    window: float = 60
 
 
 class RateLimitSettings(BaseSettings):
     enabled: bool = True
     global_limit: int = 1000
-    global_window: int = 60
+    global_window: float = 60
     endpoints: dict[str, RateLimitEndpoint] = {}
